refactor: Split Dagger component in mock/prod source set
