+++ app/src/prod/java/me/worric/souvenarius/SouvenirAppComponent.java
@@ -1,0 +1,36 @@
+package me.worric.souvenarius;
+
+import android.app.Application;
+
+import javax.inject.Singleton;
+
+import dagger.BindsInstance;
+import dagger.Component;
+import dagger.android.support.AndroidSupportInjectionModule;
+import me.worric.souvenarius.data.RoomProdModule;
+import me.worric.souvenarius.ui.FragmentContributorModule;
+import me.worric.souvenarius.ui.authwrapper.AppAuthModule;
+import me.worric.souvenarius.ui.common.ViewModelModule;
+
+@Singleton
+@Component(modules = {
+        AndroidSupportInjectionModule.class,
+        SouvenirAppModule.class,
+        ViewModelModule.class,
+        FragmentContributorModule.class,
+        AppAuthModule.class,
+        RoomProdModule.class,
+        DataModule.class
+})
+public interface SouvenirAppComponent {
+
+    @Component.Builder
+    interface Builder {
+        @BindsInstance
+        Builder application(Application application);
+        SouvenirAppComponent build();
+    }
+
+    void inject(BaseSouvenirApp app);
+
+}
