+++ app/src/prod/java/me/worric/souvenarius/data/DataModule.java
@@ -1,6 +1,5 @@
 package me.worric.souvenarius.data;
 
-import androidx.room.Room;
 import android.content.Context;
 import android.location.Geocoder;
 
@@ -16,7 +15,6 @@ import dagger.Binds;
 import dagger.Module;
 import dagger.Provides;
 import me.worric.souvenarius.R;
-import me.worric.souvenarius.data.db.AppDatabase;
 import me.worric.souvenarius.data.repository.location.LocationRepository;
 import me.worric.souvenarius.data.repository.location.LocationRepositoryImpl;
 import me.worric.souvenarius.data.repository.location.LocationTaskRunner;
@@ -36,12 +34,6 @@ import me.worric.souvenarius.di.SouvenirErrorMsgs;
 
 @Module
 public abstract class DataModule {
-
-    @Provides
-    @Singleton
-    static AppDatabase provideDatabase(@AppContext Context context) {
-        return Room.databaseBuilder(context, AppDatabase.class, AppDatabase.DB_NAME).build();
-    }
 
     @Binds
     abstract SouvenirRepository bindSouvenirRepository(SouvenirRepositoryImpl impl);
